requirements.txt
vectara-skunk-client>=0.4.14
marko==2.0.2
python-docx>=1.1.0
Jinja2==3.1.3